slot-based-collator: Refactor some internals
Move `RelayChainDataFetcher` to its own module and rename it to `RelayChainDataCache`. Also move the `core_selector` function to the `slot_based` module.
These changes are required for the 500ms blocks.
